Added Symptom interface; changed Symptom class to CommonSymptom; added idNum for CommonSymptom
